SERVICE/pick_and_place_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==40.0.2
exceptiongroup==1.3.1
fastapi==0.124.0
fastapi-health==0.4.0
flatbuffers==25.9.23
h11==0.16.0
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
Nuitka==2.8.9
numpy==1.26.4
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
python-decouple==3.4
python-multipart==0.0.5
scipy==1.15.3
six==1.17.0
spicy==0.16.0
starlette==0.50.0
sympy==1.14.0
tqdm==4.64.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
zstandard==0.25.0
